centralized-discovery-service/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
SQLAlchemy==1.4.25
psycopg2-binary==2.9.1
aiohttp==3.7.4.post0
dnspython==2.1.0
pydantic==1.8.2